Cleanup TODOs and fix Kicking Users
